Conflicts:
	test/target/classes/kr/two/bshop/ProjectController.class
	test/target/m2e-wtp/web-resources/META-INF/maven/kr.board/controller/pom.properties

+++ test/src/main/java/kr/two/entity/DiaryVO.java
@@ -19,6 +19,6 @@ public class DiaryVO {
 	private int danger;	
 	private int max;
 	private int min;
-	private int cal;
+	private float cal;
 
 }
